The library is buildable and testable
