+++ .gitignore
@@ -4,3 +4,4 @@ node_modules/
 .vscode/
 .DS_Store
 package-lock.json
+dist
