+++ package.json
@@ -3,9 +3,11 @@
   "version": "0.2.0",
   "description": "A modern SQLite store for node-cache-manager",
   "main": "dist/index.js",
+  "types": "./dist/index.d.ts",
   "type": "module",
   "scripts": {
-    "test": "nyc --reporter=html mocha"
+    "test": "node --loader ts-node/esm --test src/tests/*.ts",
+    "build": "tsc -p tsconfig.build.json"
   },
   "repository": {
     "type": "git",
@@ -30,11 +32,11 @@
   },
   "devDependencies": {
     "@types/better-sqlite3": "^7.6.2",
+    "@types/sinon": "^10.0.13",
     "cache-manager": "^5.1.3",
-    "mocha": "^10.0.0",
-    "nyc": "^15.1.0",
     "prettier": "^2.7.1",
     "sinon": "^14.0.0",
+    "ts-node": "^10.9.1",
     "typescript": "^4.9.3"
   },
   "engines": {
